# backend/public/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/categories": {
      "get": {
        "description": "獲取所有商品分類列表",
        "operationId": "CategoryController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "獲取分類列表成功"
          }
        },
        "summary": "獲取所有分類",
        "tags": [
          "商品分類管理"
        ]
      },
      "post": {
        "description": "創建新的商品分類（僅限賣家和管理員）",
        "operationId": "CategoryController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "分類創建成功"
          },
          "400": {
            "description": "請求參數錯誤"
          },
          "401": {
            "description": "未認證用戶"
          },
          "403": {
            "description": "權限不足"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "創建商品分類",
        "tags": [
          "商品分類管理"
        ]
      }
    },
...
    "/api/v1/chat/rooms": {
      "post": {
        "description": "買家和賣家之間創建或獲取聊天室",
        "operationId": "ChatController_createOrGetRoom",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "buyerId": {
                    "type": "number",
                    "description": "買家ID（賣家提供）"
                  },
                  "sellerId": {
                    "type": "number",
                    "description": "賣家ID（買家提供）"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "聊天室創建/獲取成功"
          },
          "400": {
            "description": "請求參數錯誤"
          },
          "401": {
            "description": "未認證用戶"
          },
          "403": {
            "description": "權限不足"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "創建或獲取聊天室",
        "tags": [
          "聊天系統"
        ]
      },
      "get": {
        "description": "獲取當前用戶的所有聊天室",
        "operationId": "ChatController_getUserRooms",
        "parameters": [
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每頁數量",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "頁碼",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "獲取聊天室列表成功"
          },
          "401": {
            "description": "未認證用戶"
          },
          "403": {
            "description": "權限不足"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "獲取聊天室列表",
        "tags": [
          "聊天系統"
        ]
      }
    },
    "/api/v1/chat/rooms/{roomId}/messages": {
      "get": {
        "description": "獲取指定聊天室的訊息記錄",
        "operationId": "ChatController_getMessages",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "description": "聊天室ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每頁數量",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "頁碼",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "獲取訊息成功"
          },
          "401": {
            "description": "未認證用戶"
          },
          "403": {
            "description": "權限不足或無聊天室訪問權限"
          },
          "404": {
            "description": "聊天室不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "獲取聊天訊息",
        "tags": [
          "聊天系統"
        ]
      },
      "post": {
        "description": "在指定聊天室中發送訊息",
        "operationId": "ChatController_sendMessage",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "description": "聊天室ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string",
                    "description": "訊息內容"
                  }
                },
                "required": [
                  "content"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "訊息發送成功"
          },
          "400": {
            "description": "請求參數錯誤或訊息內容為空"
          },
          "401": {
            "description": "未認證用戶"
          },
          "403": {
            "description": "權限不足或無聊天室訪問權限"
          },
          "404": {
            "description": "聊天室不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "發送訊息",
        "tags": [
          "聊天系統"
        ]
      }
    },
...
    "/api/v1/categories/{id}": {
      "get": {
        "description": "根據ID獲取指定商品分類",
        "operationId": "CategoryController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "分類ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "獲取分類成功"
          },
          "404": {
            "description": "分類不存在"
          }
        },
        "summary": "獲取單個分類",
        "tags": [
          "商品分類管理"
        ]
      },
      "patch": {
        "description": "更新指定商品分類（僅限賣家和管理員）",
        "operationId": "CategoryController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "分類ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "分類更新成功"
          },
          "400": {
            "description": "請求參數錯誤"
          },
          "401": {
            "description": "未認證用戶"
          },
          "403": {
            "description": "權限不足"
          },
          "404": {
            "description": "分類不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "更新分類",
        "tags": [
          "商品分類管理"
        ]
      },
      "delete": {
        "description": "刪除指定商品分類（僅限賣家和管理員）",
        "operationId": "CategoryController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "分類ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "分類刪除成功"
          },
          "401": {
            "description": "未認證用戶"
          },
          "403": {
            "description": "權限不足"
          },
          "404": {
            "description": "分類不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "刪除分類",
        "tags": [
          "商品分類管理"
        ]
      }
    },
...
  },
  "info": {
    "title": "Zipperoo 電商平台 API",
    "description": "Zipperoo 電商平台後端 API 文檔",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}